Save function working properly
Save function is now working properly.

NOTE: In order for save to work the node that needs to be saved requires a "Control" node which contains all of that node's data AND must be part of the "SAVE" group , for example:

The "Player" node has a child node called "Control" AND is part of the "SAVE" group.

The "dummy" node is part of the "SAVE" group but has no "Control" node and therefore it's data cannot be saved.